"format: int64" to application-crd.yaml

+++ pkg/apis/application/v1alpha1/types.go
@@ -346,7 +346,7 @@ type OperationState struct {
 	// FinishedAt contains time of operation completion
 	FinishedAt *metav1.Time `json:"finishedAt,omitempty" protobuf:"bytes,7,opt,name=finishedAt"`
 	// Try contains the number of times we have tried this sync.
-	Try int `json:"try,omitempty" protobuf:"bytes,8,opt,name=try"`
+	Try int64 `json:"try,omitempty" protobuf:"bytes,8,opt,name=try"`
 }
 
 type Info struct {
@@ -360,7 +360,7 @@ type SyncPolicy struct {
 	Automated *SyncPolicyAutomated `json:"automated,omitempty" protobuf:"bytes,1,opt,name=automated"`
 	// If an error is returned when Argo CD  tries to perform a transfer, it will retry this number of times before giving up.
 	// Setting the number to 0 makes Argo CD  do no retries (which is the default).
-	MaxRetries int `json:"maxRetries,omitempty" protobuf:"bytes,2,opt,name=maxRetries"`
+	MaxRetries int64 `json:"maxRetries,omitempty" protobuf:"bytes,2,opt,name=maxRetries"`
 }
 
 // SyncPolicyAutomated controls the behavior of an automated sync
